+++ src/file_output.jl
@@ -8,49 +8,127 @@ This function determines both for the lineplot and the csv output if
 they should be created:
     - if not, "nothing" will be returned as key list
     - if yes, the key lists of the outputs will be returned, either containing
-        - all possible keys if this is requested in the input file or
+        - all possible keys (in-/excluding flows) if this is requested in the input file or
         - only the requested keys as requested in the input file
 """
-function get_output_keys(io_settings::Dict{String,Any},
+function get_output_keys(io_settings::AbstractDict{String,Any},
                          components::Grouping)::Tuple{Union{Nothing,Vector{EnergySystems.OutputKey}},
                                                       Union{Nothing,Vector{EnergySystems.OutputKey}}}
     # determine if lineplot and csv should be created
-    do_plot_all_outputs = false
+    do_plot_all_outputs_excl_flows = false
+    do_plot_all_outputs_incl_flows = false
     do_create_plot = false
     if haskey(io_settings, "output_plot")
         do_create_plot = true  # if the key exists, set do_create_plot to true by default
-        if io_settings["output_plot"] == "all"
-            do_plot_all_outputs = true
+        if io_settings["output_plot"] == "all_excl_flows"
+            do_plot_all_outputs_excl_flows = true
+        elseif io_settings["output_plot"] == "all_incl_flows"
+            do_plot_all_outputs_incl_flows = true
         elseif io_settings["output_plot"] == "nothing"
             do_create_plot = false
+        elseif io_settings["output_plot"] == "all"
+            @error "For \"output_plot\", the input \"all\" is no longer supported. Use \"all_incl_flows\" or \"all_excl_flows\"."
+            throw(InputError)
         end
     end
 
-    do_write_all_CSV_outputs = false
+    do_write_all_CSV_outputs_excl_flows = false
+    do_write_all_CSV_outputs_incl_flows = false
     do_write_CSV = false
     if haskey(io_settings, "csv_output_keys")
         do_write_CSV = true  # if the key exists, set do_create_plot to true by default
-        if io_settings["csv_output_keys"] == "all"
-            do_write_all_CSV_outputs = true
+        if io_settings["csv_output_keys"] == "all_excl_flows"
+            do_write_all_CSV_outputs_excl_flows = true
+        elseif io_settings["csv_output_keys"] == "all_incl_flows"
+            do_write_all_CSV_outputs_incl_flows = true
         elseif io_settings["csv_output_keys"] == "nothing"
             do_write_CSV = false
+        elseif io_settings["csv_output_keys"] == "all"
+            @error "For \"csv_output_keys\", the input \"all\" is no longer supported. Use \"all_incl_flows\" or \"all_excl_flows\"."
+            throw(InputError)
         end
     end
 
+    plot_all_excluding_flows = do_plot_all_outputs_excl_flows || do_write_all_CSV_outputs_excl_flows
+    plot_all_including_flows = do_write_all_CSV_outputs_incl_flows || do_plot_all_outputs_incl_flows
+
     # collect all possible outputs of all units if needed
-    if do_plot_all_outputs || do_write_all_CSV_outputs
-        all_output_keys = Vector{EnergySystems.OutputKey}()
+    if plot_all_excluding_flows || plot_all_including_flows
+        if plot_all_excluding_flows
+            all_output_keys_excl_flows = Vector{EnergySystems.OutputKey}()
+        end
+        if plot_all_including_flows
+            all_output_keys_incl_flows = Vector{EnergySystems.OutputKey}()
+        end
         for unit in components
-            temp_dict = Dict{String,Any}(unit[2].uac => output_values(unit[2]))
-            append!(all_output_keys, output_keys(components, temp_dict))
+            output_vals = output_values(unit[2])
+            if plot_all_including_flows
+                temp_dict_incl_flows = Dict{String,Any}()
+                for output_val in output_vals
+                    if startswith(output_val, "EnergyFlow")
+                        key = String(unit[2].medium)
+                        if haskey(temp_dict_incl_flows, key)
+                            push!(temp_dict_incl_flows[key], output_val[12:end])
+                        else
+                            temp_dict_incl_flows[key] = [output_val[12:end]]
+                        end
+                    elseif startswith(output_val, "TemperatureFlow")
+                        # do nothing, temperature are added in output_keys()
+                        continue
+                    else
+                        key = unit[2].uac
+                        if haskey(temp_dict_incl_flows, key)
+                            push!(temp_dict_incl_flows[key], output_val)
+                        else
+                            temp_dict_incl_flows[key] = [output_val]
+                        end
+                    end
+                end
+                append!(all_output_keys_incl_flows, output_keys(components, temp_dict_incl_flows))
+            end
+            if plot_all_excluding_flows
+                temp_dict_excl_flows = Dict{String,Any}()
+                for output_val in output_vals
+                    if startswith(output_val, "EnergyFlow")
+                        continue
+                    elseif startswith(output_val, "TemperatureFlow")
+                        continue
+                    else
+                        key = unit[2].uac
+                        if haskey(temp_dict_excl_flows, key)
+                            push!(temp_dict_excl_flows[key], output_val)
+                        else
+                            temp_dict_excl_flows[key] = [output_val]
+                        end
+                    end
+                end
+                append!(all_output_keys_excl_flows, output_keys(components, temp_dict_excl_flows))
+            end
+        end
+        # sort keys
+        function sort_by(output_key)
+            any(startswith(output_key.value_key, p) for p in ("EnergyFlow", "TemperatureFlow")) ?
+            # flows after others: primary = medium (always present), secondary
+            lowercase("zzzzzzzzzzzzz" * string(output_key.medium) * output_key.value_key) :
+            # default: primary = unit.uac, secondary = medium (if present), tertiary
+            lowercase(string(output_key.unit.uac) * string(something(output_key.medium, "")) * output_key.value_key)
+        end
+        if plot_all_excluding_flows
+            sort!(all_output_keys_excl_flows; by=sort_by)
+        end
+        if plot_all_including_flows
+            sort!(all_output_keys_incl_flows; by=sort_by)
         end
     end
 
     # collect output keys for lineplot and csv output
     if do_create_plot  # line plot
-        output_keys_lineplot = do_plot_all_outputs ? all_output_keys : Vector{EnergySystems.OutputKey}()
-        # Collect output keys if not plotting all outputs
-        if !do_plot_all_outputs
+        if do_plot_all_outputs_incl_flows
+            output_keys_lineplot = all_output_keys_incl_flows
+        elseif do_plot_all_outputs_excl_flows
+            output_keys_lineplot = all_output_keys_excl_flows
+        else
+            output_keys_lineplot = Vector{EnergySystems.OutputKey}()
             for plot in io_settings["output_plot"]
                 key = plot[2]["key"]
                 append!(output_keys_lineplot, output_keys(components, key))
@@ -61,8 +139,10 @@ function get_output_keys(io_settings::Dict{String,Any},
     end
 
     if do_write_CSV  # csv export
-        if do_write_all_CSV_outputs # gather all outputs
-            output_keys_to_csv = all_output_keys
+        if do_write_all_CSV_outputs_incl_flows
+            output_keys_to_csv = all_output_keys_incl_flows
+        elseif do_write_all_CSV_outputs_excl_flows
+            output_keys_to_csv = all_output_keys_excl_flows
         else  # get only requested output keys from input file for CSV-export
             output_keys_to_csv = output_keys(components, io_settings["csv_output_keys"])
         end
@@ -84,7 +164,7 @@ Determines
 - the source component of each interface and [unit]
 - the target component of each interface [unit]
 
-The information is returned as single vectors with the indicees matching together.
+The information is returned as single vectors with the indices matching together.
 """
 function get_interface_information(components::Grouping)::Tuple{Int64,Vector{Any},Vector{Any},Vector{Any}}
     nr_of_interfaces = 0
@@ -94,7 +174,7 @@ function get_interface_information(components::Grouping)::Tuple{Int64,Vector{Any
     for each_component in components
         for each_outputinterface in each_component[2].output_interfaces
             medium = nothing  # reset medium
-            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Touple
+            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Tuple
                 medium = each_outputinterface[1] # then, the medium is stored separately
                 each_outputinterface = each_outputinterface[2]
             end
@@ -155,10 +235,10 @@ end
 """
 collect_interface_energies(components, nr_of_interfaces)
 
-Collects and returns the energy that was transportet through every interface.
+Collects and returns the energy that was transported through every interface.
 If the balance of an interface was not zero, the actual energy that was flowing
 is written to the outputs.
-Attention: This can lead to overfilling of demands which is currenlty not visible
+Attention: This can lead to overfilling of demands which is currently not visible
 in the sankey diagram!
 """
 function collect_interface_energies(components::Grouping, nr_of_interfaces::Int)
@@ -166,7 +246,7 @@ function collect_interface_energies(components::Grouping, nr_of_interfaces::Int)
     energies = zeros(Float64, nr_of_interfaces)
     for each_component in components
         for each_outputinterface in each_component[2].output_interfaces
-            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Touple
+            if isa(each_outputinterface, Pair) # some output_interfaces are wrapped in a Tuple
                 each_outputinterface = each_outputinterface[2]
             end
             if (isdefined(each_outputinterface, :target)
@@ -202,39 +282,80 @@ function collect_interface_energies(components::Grouping, nr_of_interfaces::Int)
 end
 
 """
-output_keys(from_config)
+output_keys(components, from_config)
 
 Transform the output keys definition in the project config file into a list of OutputKey
 items. This is done to speed up selection of values for the output in each time step,
 as this transformation has to be done only once at the beginning.
 """
-function output_keys(components::Grouping, from_config::Dict{String,Any})::Vector{EnergySystems.OutputKey}
+function output_keys(components::Grouping, from_config::AbstractDict{String,Any})::Vector{EnergySystems.OutputKey}
     outputs = Vector{EnergySystems.OutputKey}()
 
-    for unit_key in keys(from_config)
-        unit = components[unit_key]
+    all_current_media = String.(unique([bus.medium
+                                        for bus in values(components) if bus.sys_function === EnergySystems.sf_bus]))
 
-        for entry in from_config[unit_key]
-            splitted = split(String(entry))
-            if length(splitted) > 1
-                medium_key = splitted[1]
-                medium = Symbol(String(medium_key))
-                if medium in EnergySystems.medium_categories
-                    value_key = splitted[2]
+    for key in keys(from_config)
+        if key in keys(components)
+            unit_key = key
+            unit = components[unit_key]
+
+            for entry in from_config[unit_key]
+                splitted = split(String(entry), ":")
+                if length(splitted) > 1
+                    medium_key = splitted[1]
+                    medium = Symbol(String(medium_key))
+                    if medium in EnergySystems.medium_categories
+                        value_key = splitted[2]
+                    else
+                        @error "In unit \"$(unit.uac)\", the given output key \"$entry\" could not be mapped to an " *
+                               "output key. Make sure that the medium \"$(String(medium_key))\" exists in the current " *
+                               "component and that you have used \":\" as separator without any extra spaces."
+                        throw(InputError)
+                    end
                 else
-                    @error "In unit \"$(unit.uac)\", the given output key \"$entry\" could not be mapped to an output key. 
-Make sure that the medium \"$medium\" exists and that you have not used spaces accidentally. 
-Spaces are only allowed if you want to specify an output channel of a specific medium!"
+                    medium = nothing
+                    value_key = splitted[1]
+                end
+
+                push!(outputs, EnergySystems.OutputKey(; unit=unit,
+                                                       medium=medium,
+                                                       value_key=value_key))
+            end
+        elseif key in all_current_media
+            media_key = key
+            medium = Symbol(String(media_key))
+            for value_key in from_config[media_key]
+                success = false
+                in_uac, out_uac = split(value_key, "->")
+                for bus in [unit for unit in values(components) if unit.sys_function === EnergySystems.sf_bus]
+                    # consider only proxy busses or busses without proxies and busses with correct media
+                    if bus.proxy === nothing && bus.medium == medium
+                        # check if input and output exists
+                        if in_uac in keys(bus.balance_table_inputs) && out_uac in keys(bus.balance_table_outputs)
+                            output_key = "EnergyFlow " * value_key
+                            push!(outputs, EnergySystems.OutputKey(; unit=bus,
+                                                                   medium=medium,
+                                                                   value_key=output_key))
+                            output_key = "TemperatureFlow " * value_key
+                            push!(outputs, EnergySystems.OutputKey(; unit=bus,
+                                                                   medium=medium,
+                                                                   value_key=output_key))
+                            success = true
+                            break
+                        end
+                    end
+                end
+                if !success
+                    @error "The requested energy flow between components \"$(value_key)\" for medium \"$(media_key)\" could not " *
+                           "be found for the CSV output or the plot output. Note that only connections between " *
+                           "components with one or more busses but without any other component in between can be exported!"
                     throw(InputError)
                 end
-            else
-                medium = nothing
-                value_key = splitted[1]
             end
-
-            push!(outputs, EnergySystems.OutputKey(; unit=unit,
-                                                   medium=medium,
-                                                   value_key=value_key))
+        else
+            @error "The key \"$(key)\" in the provided output keys for CSV output or plot output could not be found. " *
+                   "It either has to be a medium or a component used in the current energy system."
+            throw(InputError)
         end
     end
 
@@ -268,7 +389,11 @@ function reset_file(filepath::String,
                 if outkey.medium === nothing
                     header = "$(outkey.unit.uac) $(outkey.value_key)"
                 else
-                    header = "$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key)"
+                    if startswith(outkey.value_key, "EnergyFlow") || startswith(outkey.value_key, "TemperatureFlow")
+                        header = "$(outkey.medium) $(outkey.value_key)"
+                    else
+                        header = "$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key)"
+                    end
                 end
                 write(file_handle, ";$header")
             end
@@ -333,7 +458,7 @@ Dump a bunch of information to file that might be useful to explain the result o
 This is mostly used for debugging and development purposes, but might prove useful in
 general to find out why the energy system behaves in the simulation as it does.
 """
-function dump_auxiliary_outputs(project_config::Dict{AbstractString,Any},
+function dump_auxiliary_outputs(project_config::AbstractDict{AbstractString,Any},
                                 components::Grouping,
                                 order_of_operations::StepInstructions,
                                 sim_params::Dict{String,Any})
@@ -397,9 +522,10 @@ function create_profile_line_plots(outputs_plot_data::Union{Nothing,Matrix{Float
                                    outputs_plot_keys::Union{Nothing,Vector{EnergySystems.OutputKey}},
                                    outputs_plot_weather::Union{Nothing,Matrix{Float64}},
                                    outputs_plot_weather_keys::Union{Nothing,Vector{String}},
-                                   project_config::Dict{AbstractString,Any},
+                                   project_config::AbstractDict{AbstractString,Any},
                                    sim_params::Dict{String,Any})
-    plot_all = project_config["io_settings"]["output_plot"] == "all"
+    plot_all = isa(project_config["io_settings"]["output_plot"], String) &&
+               project_config["io_settings"]["output_plot"][1:3] == "all"
     plot_data = outputs_plot_data !== nothing
     plot_weather = outputs_plot_weather !== nothing
 
@@ -410,6 +536,8 @@ function create_profile_line_plots(outputs_plot_data::Union{Nothing,Matrix{Float
         for outkey in outputs_plot_keys
             if outkey.medium === nothing
                 push!(labels, string("$(outkey.unit.uac) $(outkey.value_key)"))
+            elseif startswith(outkey.value_key, "EnergyFlow") || startswith(outkey.value_key, "TemperatureFlow")
+                push!(labels, string("$(outkey.medium) $(outkey.value_key)"))
             else
                 push!(labels, string("$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key)"))
             end
@@ -426,9 +554,40 @@ function create_profile_line_plots(outputs_plot_data::Union{Nothing,Matrix{Float
         scale_fact = Float64[]
         if plot_data
             for plot in project_config["io_settings"]["output_plot"]
-                push!(axis, string(plot[2]["axis"]))
-                push!(unit, string(plot[2]["unit"]))
-                push!(scale_fact, plot[2]["scale_factor"])
+                if occursin("->", first(plot[2]["key"])[2][1])
+                    # Here we are dealing with EnergyFlow and TemperatureFlow --> Two meta information required if 
+                    # temperature should be plotted
+                    if !isa(plot[2]["axis"], AbstractVector) || length(plot[2]["axis"]) !== 2 ||
+                       !isa(plot[2]["unit"], AbstractVector) || length(plot[2]["unit"]) !== 2 ||
+                       !isa(plot[2]["scale_factor"], AbstractVector) || length(plot[2]["scale_factor"]) !== 2
+                        # only one set of meta information given. Use the given one both for energy and temperature flow
+                        push!(axis, isa(plot[2]["axis"], AbstractVector) ? plot[2]["axis"][1] : plot[2]["axis"])
+                        push!(unit, isa(plot[2]["unit"], AbstractVector) ? plot[2]["unit"][1] : plot[2]["unit"])
+                        push!(scale_fact,
+                              isa(plot[2]["scale_factor"], AbstractVector) ? plot[2]["scale_factor"][1] :
+                              plot[2]["scale_factor"])
+
+                        push!(axis, "nothing")
+                        push!(unit, "nothing")
+                        push!(scale_fact, NaN)
+                        @info "For the generation of the output_plot, the meta information for entry $(plot[1]) " *
+                              "do not contain two values. Therefore, only energy values will be output. If you want " *
+                              "to output also a corresponding temperature, provide two meta information: " *
+                              "[EnergyFlow, TemperatureFlow] for axis, unit and scale_factor."
+                    else
+                        push!(axis, string(plot[2]["axis"][1]))
+                        push!(unit, string(plot[2]["unit"][1]))
+                        push!(scale_fact, plot[2]["scale_factor"][1])
+
+                        push!(axis, string(plot[2]["axis"][2]))
+                        push!(unit, string(plot[2]["unit"][2]))
+                        push!(scale_fact, plot[2]["scale_factor"][2])
+                    end
+                else
+                    push!(axis, string(plot[2]["axis"]))
+                    push!(unit, string(plot[2]["unit"]))
+                    push!(scale_fact, plot[2]["scale_factor"])
+                end
             end
         end
         if plot_weather
@@ -445,6 +604,8 @@ function create_profile_line_plots(outputs_plot_data::Union{Nothing,Matrix{Float
             for outkey in outputs_plot_keys
                 if outkey.medium === nothing
                     push!(labels, string("$(outkey.unit.uac) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
+                elseif startswith(outkey.value_key, "EnergyFlow") || startswith(outkey.value_key, "TemperatureFlow")
+                    push!(labels, string("$(outkey.medium) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
                 else
                     push!(labels,
                           string("$(outkey.unit.uac) $(outkey.medium) $(outkey.value_key) [$(unit[n])] ($(axis[n]))"))
@@ -465,6 +626,33 @@ function create_profile_line_plots(outputs_plot_data::Union{Nothing,Matrix{Float
     else
         time_x = outputs_plot_weather[:, 1]
     end
+    # filter NaN values
+    if plot_data
+        idxs_to_remove = Int[]
+        for col in axes(outputs_plot_data, 2)
+            if col == 1 # never remove date
+                continue
+            elseif all(isnan, outputs_plot_data[:, col])
+                push!(idxs_to_remove, col)
+            end
+        end
+        if !plot_all
+            for (idx, ax) in enumerate(axis)
+                if ax == "nothing"
+                    push!(idxs_to_remove, idx + 1)
+                end
+            end
+            idxs_to_remove = unique(idxs_to_remove)
+        end
+        if !isempty(idxs_to_remove)
+            outputs_plot_data = outputs_plot_data[:, setdiff(axes(outputs_plot_data, 2), idxs_to_remove)]
+            labels = labels[setdiff(1:length(labels), idxs_to_remove .- 1)]
+            if !plot_all
+                axis = axis[setdiff(1:length(axis), idxs_to_remove .- 1)]
+                scale_fact = scale_fact[setdiff(1:length(scale_fact), idxs_to_remove .- 1)]
+            end
+        end
+    end
     output_plot_time_unit = haskey(project_config["io_settings"], "output_plot_time_unit") ?
                             project_config["io_settings"]["output_plot_time_unit"] : "date"
     if !(output_plot_time_unit in ["seconds", "minutes", "hours", "date"])
@@ -481,7 +669,7 @@ function create_profile_line_plots(outputs_plot_data::Union{Nothing,Matrix{Float
     else
         if sim_params["start_date"] === nothing
             start_date = Dates.DateTime("2015/1/1 00:00:00", "yyyy/m/d HH:MM:SS")
-            @info ("Date of first data point in ouput line plot is set to 01-01-2015 00:00:00, as the simulation start time is not given as date.")
+            @info ("Date of first data point in output line plot is set to 01-01-2015 00:00:00, as the simulation start time is not given as date.")
         else
             start_date = sim_params["start_date"]
         end
@@ -544,14 +732,14 @@ Inputs:
 output_all_sourcenames and *sinknames are vectors with names of the source and sink of each interface
 output_all_values are logs with data from each timestep in the shape [timestep,interface].
 medium_of_interface is a vector of the medium corresponding to each interface
-nr_of_interfaces is the total number of iterfaces in the current energy system
+nr_of_interfaces is the total number of interfaces in the current energy system
 """
 function create_sankey(output_all_sourcenames::Vector{Any},
                        output_all_targetnames::Vector{Any},
                        output_all_values::Matrix{Float64},
                        medium_of_interfaces::Vector{Any},
                        nr_of_interfaces::Int64,
-                       io_settings::Dict{String,Any})
+                       io_settings::AbstractDict{String,Any})
 
     # sum up data of each interface
     output_all_value_sum = zeros(Float64, nr_of_interfaces)
@@ -613,7 +801,7 @@ function create_sankey(output_all_sourcenames::Vector{Any},
     medium_labels = [split(string(s), '.')[end] for s in medium_of_interfaces]
     unique_medium_labels = unique(medium_labels)
 
-    if io_settings["sankey_plot"] == "default" || !isa(io_settings["sankey_plot"], Dict{})
+    if io_settings["sankey_plot"] == "default" || !isa(io_settings["sankey_plot"], AbstractDict{})
         if length(unique_medium_labels) > 1
             colors = get(ColorSchemes.roma,
                          (0:(length(unique_medium_labels) - 1)) ./ (length(unique_medium_labels) - 1))
@@ -629,7 +817,7 @@ function create_sankey(output_all_sourcenames::Vector{Any},
                 color_map[medium] = RGB(color_entry[1] / 255, color_entry[2] / 255, color_entry[3] / 255)
             catch e
                 @error "The color for the sankey '$color' of medium '$medium' could not be detected. " *
-                       "The following error occured: $e\n" *
+                       "The following error occurred: $e\n" *
                        "Color has to be one of: $(collect(keys(Colors.color_names)))"
                 throw(InputError)
             end